core/requirements.txt
groq>=0.4.0
colorama>=0.4.6
python-dotenv>=1.0.0
